backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
google-genai==0.5.0
pytest==8.3.4
httpx==0.28.1
